feat: restructure code editor component
