+++ web/src/components/features/workspace/CodeEditor/autocomplete/snippets.ts
